created LoadingScreen component and refactored loading screens to use it also changed logic for timing on LoadingScreen
